Move formgrader tests into a subdirectory
