+++ nbgrader/tests/test_formgrader/files
@@ -1,0 +1,1 @@
+../../../docs/user_guide/example
